Split HtmlDocument to component and renderer
